Use explicit mime type when creating parts from file
Mime type guessing works well if the file has the right extension
but not so if using a tmp file for example.

This results in data part sent as an octet stream, which is fine by itself,
but Everial API has a hard time processing it if the content is not an image.
It will through a 500 for any content other than an image sent as an
octet stream.

To prevent this we force the content type of the data part with the
value detected on the file's content.